Create multidb base class definition
This is to simplify admin database router definition in applications
using legacy/external databases
